Use one thread per exchange
- Showing threads only for stopped message. It would bloats the UI too
much on each running message and surely make it flickering.
- Call Starting/exiting Threads when breakpoint hit, resume all. Threads
id are not reused but this is an arbitrary number anyway. The name is
more important for the end-user perspective, it is the exchange ID.
Cannot use the Exchange ID as thread id because the threadid in Debug
Adapter protocol specification is a number and Exchange ID is a string.
- it is important for Eclipse client to call the thread start and stop,
otherwise, it is not clearing the list of threds even after a new call
on threads() method

Signed-off-by: Aurélien Pupier <apupier@redhat.com>