FUSETOOLS2-1458 - Sign jar on Main branch
- 2 secrets have been provided on GitHub camel-debug-adapter repository:
MAVEN_GPG_PRIVATE_KEY and MAVEN_GPG_PASSPHRASE . They are mine (apupier)
gpg keys and passphrase. The private key must be provided including the
"delimiters" `-----BEGIN PGP PRIVATE KEY BLOCK-----` and `-----END PGP
PRIVATE KEY BLOCK-----`
- the signing is done in verify phase so modified job for it
- the signing requires to be on a branch from the repository as it is
using secrets so cannot be tested on a normal Pull Request. That's the
reason of the profile to sign.
- doing the Sonar and Signing in the same Github Action flow to avoid
rebuilding and replaying test on the same configuration. The deploy will
be added in the same job too in next phase.
- note: no need of pinentry loopback configuration as maven-gpg-plugin
3.0.1 is used
- most useful documentation:
https://central.sonatype.org/publish/requirements/gpg/ and
https://github.com/actions/setup-java/blob/main/docs/advanced-usage.md#publishing-using-apache-maven


Signed-off-by: Aurélien Pupier <apupier@redhat.com>