FUSETOOLS2-1433 - resume a single thread
It comes with a refactor to have a better scoping of variables. It
really ease reading code as the scope per Exchange becomes important to
be able to deal with as one Exchange is a thread. It will be slower in
terms of performance but it should not have a visible impact unless
there are a lot of breakpoints at the same time, which will anyway won't
be very readable in UI

Signed-off-by: Aurélien Pupier <apupier@redhat.com>