WIP FUSETOOLS2-1624 - Support automatic restart
When Automatic restart of Camel route is configured, the JMX connection
vanishes for few seconds. Adding code to be able to reconnect. As a
first iteration ,allow 5 seconds for the connection to be up again.

- how to test it?
- How to check we have no potential zombies if reconnection last more
than 5 seconds or if it was an abrupt stop?
- if automatic reload is done in the one second time between two check
of suspended nodes, we won't detect the reload. For now, I do not see a
better way than to live with it as it shouldn't occur a lot, with a very
simple use case, it is taking always more than a second on my relatively
fast machine.

Current proposal is working only with Quarkus with a not robust
implementation and I cannot manage to write a test for this case.

i tis not working with Camel JBangand I do not see how to detect that
there was a reload. I'm wondering if the breakpoints are kept on route
reload, it would not be enough to work on most cases.

Maybe a better approach is to support the "restart" of the debug adapter
which is available in VS Code UI when the code has been detected as
modified.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>