FUSETOOLS2-1819 - Fix regression when updating value
it was introduced because new version of the protocol has more
"laziness". The variables are not all calculated as before
automatically.

To provide a non-regression test, the DummyCamelDebugClient used in test
must be drastically updated, along with all tests. I feel like it will
easier to have a UI test to check this use case. It will detect issue
later but ensure to follow the real current implementation of the
client.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>